Add files via upload
Genex coin is a new payment ecosystem that deploys a digital currency and makes innovative use of blockchain technology Digital currency or cryptocurrency is a form of
electronic money. Most digital currencies are designed to be more secure, remove intermediaries and be anonymous. We LOVE NFTS and we want artists to be rewarded fairly for their hard work. However, we also understand the need to create a self-sustaining, self-incentivizing and evergrowing ecosystem. That is why we have created Genex coin.